Split CFG into a model and a view - step 2.
