+++ angr/analyses/cfg/__init__.py
@@ -9,4 +9,4 @@ from .cfg_fast_soot import CFGFastSoot
 # things to make your life easier
 from .cfg_arch_options import CFGArchOptions
 from .cfg_utils import CFGUtils
-from .cfg_node import CFGNode
+from ...knowledge_plugins.cfg import CFGNode
