+++ angr/knowledge_plugins/cfg/__init__.py
@@ -1,3 +1,5 @@
 
+from .cfg_node import CFGNode, CFGENode
+from .indirect_jump import IndirectJump
 from .cfg_model import CFGModel
 from .cfg_manager import CFGManager
