+++ angr/knowledge_plugins/cfg/cfg_model.py
@@ -1,11 +1,15 @@
 
 import networkx
 
+from ...serializable import Serializable
 
-class CFGModel:
+
+class CFGModel(Serializable):
     """
     This class describes a Control Flow Graph for a specific range of code.
     """
+
+    __slots__ = ('graph', 'jump_tables', 'memory_data', 'insn_addr_to_memory_data', )
 
     def __init__(self):
 
@@ -20,3 +24,9 @@ class CFGModel:
         self.memory_data = { }
         # A mapping between address of the instruction that's referencing the memory data and the memory data itself
         self.insn_addr_to_memory_data = { }
+
+    #
+    # CFG View
+    #
+
+
